+++ src/main/java/com/ats/tril/model/Item.java
@@ -13,14 +13,14 @@ public class Item {
 	private int grpId; 
 	private int subGrpId;
 	private Float itemOpRate; 
-	private int itemOpQty; 
+	private Float itemOpQty; 
 	private Float itemClRate; 
-	private int itemClQty; 
+	private Float itemClQty; 
 	private String itemDate; 
 	private Float itemWt; 
-	private int itemMinLevel; 
-	private int itemMaxLevel; 
-	private int itemRodLevel; 
+	private Float itemMinLevel; 
+	private Float itemMaxLevel; 
+	private Float itemRodLevel; 
 	private String itemLocation; 
 	private String itemAbc; 
 	private int itemIsCritical; 
@@ -98,10 +98,10 @@ public class Item {
 	public void setItemOpRate(Float itemOpRate) {
 		this.itemOpRate = itemOpRate;
 	}
-	public int getItemOpQty() {
+	public Float getItemOpQty() {
 		return itemOpQty;
 	}
-	public void setItemOpQty(int itemOpQty) {
+	public void setItemOpQty(Float itemOpQty) {
 		this.itemOpQty = itemOpQty;
 	}
 	public Float getItemClRate() {
@@ -110,10 +110,10 @@ public class Item {
 	public void setItemClRate(Float itemClRate) {
 		this.itemClRate = itemClRate;
 	}
-	public int getItemClQty() {
+	public Float getItemClQty() {
 		return itemClQty;
 	}
-	public void setItemClQty(int itemClQty) {
+	public void setItemClQty(Float itemClQty) {
 		this.itemClQty = itemClQty;
 	}
 	public String getItemDate() {
@@ -128,22 +128,22 @@ public class Item {
 	public void setItemWt(Float itemWt) {
 		this.itemWt = itemWt;
 	}
-	public int getItemMinLevel() {
+	public Float getItemMinLevel() {
 		return itemMinLevel;
 	}
-	public void setItemMinLevel(int itemMinLevel) {
+	public void setItemMinLevel(Float itemMinLevel) {
 		this.itemMinLevel = itemMinLevel;
 	}
-	public int getItemMaxLevel() {
+	public Float getItemMaxLevel() {
 		return itemMaxLevel;
 	}
-	public void setItemMaxLevel(int itemMaxLevel) {
+	public void setItemMaxLevel(Float itemMaxLevel) {
 		this.itemMaxLevel = itemMaxLevel;
 	}
-	public int getItemRodLevel() {
+	public Float getItemRodLevel() {
 		return itemRodLevel;
 	}
-	public void setItemRodLevel(int itemRodLevel) {
+	public void setItemRodLevel(Float itemRodLevel) {
 		this.itemRodLevel = itemRodLevel;
 	}
 	public String getItemLocation() {
